+++ src/solidmodels/solidmodels.jl
@@ -67,6 +67,8 @@ import DeviceLayout:
     level,
     norm,
     render!,
+    round_to_curvilinearpolygon,
+    rounded_corner_segment,
     uniquename,
     isapprox_angle
 import DeviceLayout.Paths: trace, gap, offset, extent, pathlength, bspline_approximation